tests/wpt/web-platform-tests/css/css-logical/parsing/margin-block-inline-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Logical Properties and Values: parsing margin-block and margin-inline with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-logical/#propdef-margin-block">
<meta name="assert" content="margin-block, margin-inline support the full grammar '<'margin-top'>{1,2}'.">
<meta name="assert" content="margin-block, margin-inline longhands support the full grammar '<'margin-top'>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("margin-block-start", "calc(20% + 10px)");
test_valid_value("margin-block-start", "auto");
test_valid_value("margin-block-end", "-10px");
test_valid_value("margin-inline-start", "-20%");
test_valid_value("margin-inline-end", "calc(2em + 3ex)");

test_valid_value("margin-block", "auto");
test_valid_value("margin-block", "-10px");
test_valid_value("margin-block", "calc(2em + 3ex) -20%");
test_valid_value("margin-block", "auto auto", "auto");
test_valid_value("margin-block", "-20% calc(20% + 10px)");
test_valid_value("margin-inline", "20%");
test_valid_value("margin-inline", "calc(2em + 3ex)");
test_valid_value("margin-inline", "-10px auto");
test_valid_value("margin-inline", "auto calc(2em + 3ex)");
</script>
</body>
</html>
